Changes required for rabbitmq 7.0.0

+++ src/Queues/RabbitMq/src/RabbitMqClientOptions.cs
@@ -65,7 +65,7 @@ public class RabbitMqClientOptions : IOptions<RabbitMqClientOptions>
     /// </summary>
     /// <remarks>For concurrency greater than one this removes the guarantee that consumers handle messages in the order they receive them.
     /// In addition to that consumers need to be thread/concurrency safe.</remarks>
-    public int ConsumerDispatchConcurrency { get; set; } = 1;
+    public ushort ConsumerDispatchConcurrency { get; set; } = 1;
 
     public RabbitMqClientOptions Value => this;
 }
